requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
playwright==1.49.1
lxml==5.3.0
langchainhub==0.1.21
python-dotenv==1.0.1
